Fixes for compatibility with future breaking changes of Analyzer.
Mostly to access elements corresponding to referenced functions
using API that was added for this purpose. Even now function
types not always have elements, and we want to move Analyzer to
use structural FunctionType(s) as much as possible.

For FunctionType(s) that are instantiations of typedef(s) we
still keep the element and type arguments, because, as we
discussed earlier, Angular is not ready to write generic
function types.

PiperOrigin-RevId: 271177692