Reduce code size of generated host views
This changes HostView.injector to be initialized in HostView.create() rather
than the constructor. While the constructor is theoretically preferable, as it
allows the injector field to be final, it results in more generated code. This
is because having a constructor parameter requires generating a constructor for
every implementation of HostView to forward an injector parameter, whereas
HostView.create() is defined once.

PiperOrigin-RevId: 273808556