Remove reliance on SafeValue.toString()
There is no legitimate need for SafeValue.toString() to return the sanitized value,
as it should only be consumed by the sanitizer which has access to the internal
fields.

The removed test cases were testing an implementation detail rather than the
behavior that matters: a value passed to a bypassSecurityTrust* method is
trusted by the sanitizer. This case is handled in the test cases above.

PiperOrigin-RevId: 274619989